hide the cursor when repainting
There is another possible source of the annoying flicker effect we will take care of now.
It’s possible that the cursor might be displayed in the middle of the screen somewhere for a split second while the terminal is drawing to the screen.
To make sure that doesn’t happen, let’s hide the cursor before refreshing the screen, and show it again immediately after the refresh finishes.