backspace
In edit mode, I want to use the backspace key to remove the character to the left of the cursor.
If the cursor is at the beginning of the line, the entire line should be appended to the line above it.
If the cursor is in the top left corner, nothing should happen.